officetel_pipeline/independent/requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
requests==2.31.0
urllib3==1.26.15
